+++ package.json
@@ -20,6 +20,7 @@
   "homepage": "https://github.com/Elenapark/NodeJS_Basic#readme",
   "dependencies": {
     "date-fns": "^2.29.3",
+    "express": "^4.18.1",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
